fixed the components for the shopping cart
