+++ components/Cart.js
@@ -1,65 +1,195 @@
 import React from 'react'
 import Link from 'next/link'
 
-const Cart = () => {
-  const [cartOpen, setCartOpen] = React.useState(false)
+const Checkout = () => {
   return (
-    <div className="p-5">
-      <div className="relative">
-        <div className="absolute w-full  rounded-b border-t-0 z-10">
-          <div className="shadow-xl w-64">
-            <div className="p-2 flex bg-white hover:bg-gray-100 cursor-pointer border-b border-gray-100">
-              {/* put image at the div below */}
-              <div className="p-2 w-12"></div>
-              <div className="flex-auto text-sm w-32">
-                <div className="font-bold">Product 1</div>
-                <div className="truncate">Product 1 description</div>
-                <div className="text-gray-400">Qt: 2</div>
-              </div>
-              <div className="flex flex-col w-18 font-medium items-end">
-                <div class="w-4 h-4 mb-6 hover:bg-red-200 rounded-full cursor-pointer text-red-700">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="100%"
-                    height="100%"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                    stroke-width="2"
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    class="feather feather-trash-2 "
-                  >
-                    <polyline points="3 6 5 6 21 6"></polyline>
-                    <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path>
-                    <line x1="10" y1="11" x2="10" y2="17"></line>
-                    <line x1="14" y1="11" x2="14" y2="17"></line>
-                  </svg>
+    <div className="flex justify-center my-6">
+      <div className="flex flex-col w-full p-8 text-gray-800 bg-white shadow-lg pin-r pin-y md:w-4/5 lg:w-4/5">
+        <div className="flex-1">
+          <div className="flex justify-between border-b pb-8">
+            <h1 class="font-semibold text-2xl">Shopping Cart</h1>
+            <h2 class="font-semibold text-2xl">3 Items</h2>
+          </div>
+          <table className="w-full text-sm lg:text-base" cellspacing="0">
+            <thead>
+              <tr className="h-12 uppercase">
+                <th className="hidden md:table-cell"></th>
+                <th className="text-left">Product</th>
+                <th className="lg:text-right text-left pl-5 lg:pl-0">
+                  <span className="lg:hidden" title="Quantity">
+                    Qtd
+                  </span>
+                  <span className="hidden lg:inline">Quantity</span>
+                </th>
+                <th className="hidden text-right md:table-cell">Unit price</th>
+                <th className="text-right">Total price</th>
+              </tr>
+            </thead>
+            <tbody>
+              <td class="hidden pb-4 md:table-cell">
+                <a href="#">
+                  {/* <img src="https://limg.app/i/Calm-Cormorant-Catholic-Pinball-Blaster-yM4oub.jpeg" class="w-20 rounded" alt="Thumbnail"> */}
+                </a>
+              </td>
+              <td>
+                <a>
+                  <p class="mb-2 md:ml-4">Smocked Top</p>
+                  <form action="" method="POST">
+                    <button type="submit" class="text-gray-700 md:ml-4">
+                      <small>(Remove item)</small>
+                    </button>
+                  </form>
+                </a>
+              </td>
+              <td class="justify-center md:justify-end md:flex mt-6">
+                <div class="w-20 h-10">
+                  <div class="relative flex flex-row w-full h-8">
+                    <input
+                      type="number"
+                      value="2"
+                      class="w-full font-semibold text-center text-gray-700 bg-gray-200 outline-none focus:outline-none hover:text-black focus:text-black"
+                    />
+                  </div>
                 </div>
-                P299
+              </td>
+              <td class="hidden text-right md:table-cell">
+                <span class="text-sm lg:text-base font-medium">P10.00</span>
+              </td>
+              <td class="text-right">
+                <span class="text-sm lg:text-base font-medium">P20.00</span>
+              </td>
+            </tbody>
+          </table>
+          <hr className="pb-6 mt-6" />
+          <div className="my-4 mt-6 -mx-2 lg:flex">
+            <div className="lg:px-2 lg:w-1/2">
+              <div className="p-4 bg-gray-100 rounded-full">
+                <h1 className="ml-2 font-bold uppercase">Coupon Code</h1>
               </div>
+              <div className="p-4">
+                <p className="mb-4 italic">If you have a coupon code, please enter it in the box below</p>
+                <div className="justify-center md:flex">
+                  <form action="" method="POST">
+                    <div className="flex items-center w-full h-13 pl-3 bg-white bg-gray-100 border rounded-full">
+                      <input
+                        type="coupon"
+                        name="code"
+                        id="coupon"
+                        placeholder="Apply coupon"
+                        value="90off"
+                        class="w-full bg-gray-100 outline-none appearance-none focus:outline-none active:outline-none"
+                      />
+                      <button
+                        type="submit"
+                        className="text-sm flex items-center px-3 py-1 text-white bg-gray-800 rounded-full outline-none md:px-4 hover:bg-gray-700 focus:outline-none active:outline-none"
+                      >
+                        <svg
+                          aria-hidden="true"
+                          data-prefix="fas"
+                          data-icon="gift"
+                          class="w-8"
+                          xmlns="http://www.w3.org/2000/svg"
+                          viewBox="0 0 512 512"
+                        >
+                          <path
+                            fill="currentColor"
+                            d="M32 448c0 17.7 14.3 32 32 32h160V320H32v128zm256 32h160c17.7 0 32-14.3 32-32V320H288v160zm192-320h-42.1c6.2-12.1 10.1-25.5 10.1-40 0-48.5-39.5-88-88-88-41.6 0-68.5 21.3-103 68.3-34.5-47-61.4-68.3-103-68.3-48.5 0-88 39.5-88 88 0 14.5 3.8 27.9 10.1 40H32c-17.7 0-32 14.3-32 32v80c0 8.8 7.2 16 16 16h480c8.8 0 16-7.2 16-16v-80c0-17.7-14.3-32-32-32zm-326.1 0c-22.1 0-40-17.9-40-40s17.9-40 40-40c19.9 0 34.6 3.3 86.1 80h-86.1zm206.1 0h-86.1c51.4-76.5 65.7-80 86.1-80 22.1 0 40 17.9 40 40s-17.9 40-40 40z"
+                          />
+                        </svg>
+                        <span className="font-medium">Apply coupon</span>
+                      </button>
+                    </div>
+                  </form>
+                </div>
+              </div>
+              <div className="p-4 mt-6 bg-gray-100 rounded-full">
+                <h1 className="ml-2 font-bold uppercase">Instruction for seller</h1>
+              </div>
+              <div className="p-4">
+                <p className="mb-4 italic">
+                  If you have some information for the seller you can leave them in the box below
+                </p>
+                <textarea className="w-full h-24 p-2 bg-gray-100 rounded"></textarea>
+              </div>
+              <a href="#" class="flex font-semibold text-indigo-600 text-sm mt-10">
+                <svg class="fill-current mr-2 text-indigo-600 w-4" viewBox="0 0 448 512">
+                  <path d="M134.059 296H436c6.627 0 12-5.373 12-12v-56c0-6.627-5.373-12-12-12H134.059v-46.059c0-21.382-25.851-32.09-40.971-16.971L7.029 239.029c-9.373 9.373-9.373 24.569 0 33.941l86.059 86.059c15.119 15.119 40.971 4.411 40.971-16.971V296z" />
+                </svg>
+                Continue Shopping
+              </a>
             </div>
-            <div class="p-4 justify-center flex">
-              <Link href = '/checkout'>
-                <button
-                  class="text-base  undefined  hover:scale-110 focus:outline-none flex justify-center px-4 py-2 rounded font-bold cursor-pointer 
-        hover:bg-teal-700 hover:text-teal-100 
-        bg-teal-100 
-        text-teal-700 
-        border duration-200 ease-in-out 
-        border-teal-600 transition"
-                >
-                  Checkout 
-                  <span> P299</span>
-                </button>
-              </Link>
+            <div className="lg:px-2 lg:w-1/2">
+              <div className="p-4 bg-gray-100 rounded-full">
+                <h1 className="ml-2 font-bold uppercase">Order Details</h1>
+              </div>
+              <div className="p-4">
+                <p class="mb-6 italic">
+                  Shipping and additionnal costs are calculated based on values you have entered
+                </p>
+                <div class="flex justify-between border-b">
+                  <div class="lg:px-4 lg:py-2 m-2 text-lg lg:text-xl font-bold text-center text-gray-800">Subtotal</div>
+                  <div class="lg:px-4 lg:py-2 m-2 lg:text-lg font-bold text-center text-gray-900">P499</div>
+                </div>
+                <div class="flex justify-between pt-4 border-b">
+                  <div class="flex lg:px-4 lg:py-2 m-2 text-lg lg:text-xl font-bold text-gray-800">
+                    <form action="" method="POST">
+                      <button type="submit" class="mr-2 mt-1 lg:mt-2">
+                        <svg
+                          aria-hidden="true"
+                          data-prefix="far"
+                          data-icon="trash-alt"
+                          class="w-4 text-red-600 hover:text-red-800"
+                          xmlns="http://www.w3.org/2000/svg"
+                          viewBox="0 0 448 512"
+                        >
+                          <path
+                            fill="currentColor"
+                            d="M268 416h24a12 12 0 0012-12V188a12 12 0 00-12-12h-24a12 12 0 00-12 12v216a12 12 0 0012 12zM432 80h-82.41l-34-56.7A48 48 0 00274.41 0H173.59a48 48 0 00-41.16 23.3L98.41 80H16A16 16 0 000 96v16a16 16 0 0016 16h16v336a48 48 0 0048 48h288a48 48 0 0048-48V128h16a16 16 0 0016-16V96a16 16 0 00-16-16zM171.84 50.91A6 6 0 01177 48h94a6 6 0 015.15 2.91L293.61 80H154.39zM368 464H80V128h288zm-212-48h24a12 12 0 0012-12V188a12 12 0 00-12-12h-24a12 12 0 00-12 12v216a12 12 0 0012 12z"
+                          />
+                        </svg>
+                      </button>
+                    </form>
+                    Coupon "90off"
+                  </div>
+                  <div className="lg:px-4 lg:py-2 m-2 lg:text-lg font-bold text-center text-green-700">-P150</div>
+                </div>
+                <div className="flex justify-between pt-4 border-b">
+                  <div className="lg:px-4 lg:py-2 m-2 text-lg lg:text-xl font-bold text-center text-gray-800">
+                    New Subtotal
+                  </div>
+                  <div className="lg:px-4 lg:py-2 m-2 lg:text-lg font-bold text-center text-gray-900">P250</div>
+                </div>
+                <div className="flex justify-between pt-4 border-b">
+                  <div className="lg:px-4 lg:py-2 m-2 text-lg lg:text-xl font-bold text-center text-gray-800">
+                    Total
+                  </div>
+                  <div className="lg:px-4 lg:py-2 m-2 lg:text-lg font-bold text-center text-gray-900">P200</div>
+                </div>
+                <a href="#">
+                  <button className="flex justify-center w-full px-10 py-3 mt-6 font-medium text-white uppercase bg-gray-800 rounded-full shadow item-center hover:bg-gray-700 focus:shadow-outline focus:outline-none">
+                    <svg
+                      aria-hidden="true"
+                      data-prefix="far"
+                      data-icon="credit-card"
+                      class="w-8"
+                      xmlns="http://www.w3.org/2000/svg"
+                      viewBox="0 0 576 512"
+                    >
+                      <path
+                        fill="currentColor"
+                        d="M527.9 32H48.1C21.5 32 0 53.5 0 80v352c0 26.5 21.5 48 48.1 48h479.8c26.6 0 48.1-21.5 48.1-48V80c0-26.5-21.5-48-48.1-48zM54.1 80h467.8c3.3 0 6 2.7 6 6v42H48.1V86c0-3.3 2.7-6 6-6zm467.8 352H54.1c-3.3 0-6-2.7-6-6V256h479.8v170c0 3.3-2.7 6-6 6zM192 332v40c0 6.6-5.4 12-12 12h-72c-6.6 0-12-5.4-12-12v-40c0-6.6 5.4-12 12-12h72c6.6 0 12 5.4 12 12zm192 0v40c0 6.6-5.4 12-12 12H236c-6.6 0-12-5.4-12-12v-40c0-6.6 5.4-12 12-12h136c6.6 0 12 5.4 12 12z"
+                      />
+                    </svg>
+                    <span className="ml-2 mt-5px">Procceed to checkout</span>
+                  </button>
+                </a>
+              </div>
             </div>
           </div>
         </div>
-        <div class="h-32"></div>
       </div>
     </div>
   )
 }
 
-export default Cart
+export default Checkout
